v2, refactor for ease of use
Refactored the widget to better fit the spirit of pluggable widgets. Instead of creating Mendix objects and calling microflows directly in the widget, we set an attribute on the context object when the captcha process is complete. Then the token can be validated on the server upon call of a microflow, where presumably other validation is done as well (like data in a form, or username/password for login).